+++ client-app/app/pages/editor/leftPanel/spriteSheets/spriteSheets.ts
@@ -1,16 +1,17 @@
-import BaseComponent from "../../../../baseComponent";
+
 
 declare const RF;
 
+import {BaseComponent} from "../../../../baseComponent";
 import './spriteSheets.scss'
-import SpriteSheet from "../../../../../engine/model/impl/spriteSheet";
+import {SpriteSheet} from "../../../../../engine/model/impl/spriteSheet";
 import {alertEx} from "../../../../providers/userDefinedFns";
 
 @RF.decorateComponent({
     name: 'app-sprite-sheets',
     template: require('./spriteSheets.html')
 })
-export default class SpriteSheets extends BaseComponent {
+export class SpriteSheets extends BaseComponent {
     constructor(){
         super();
     }
@@ -22,15 +23,15 @@ export default class SpriteSheets extends BaseComponent {
         this.editData.currSpriteSheetInEdit = sprSh.clone();
         RF.getComponentById('spriteSheetDialog').open();
     }
-    deleteSpriteSheet(model){
+    async deleteSpriteSheet(model){
         let hasDepends = this.editData.game.repository.getArray('GameObject').filter(it=>it.spriteSheet.id==model.id).length>0;
         if (hasDepends) {
             alertEx(this.i18n.canNotDelete(model));
             return;
         }
-        this.utils.deleteModel(model,()=>{
-            this.restFileSystem.removeFile(model.resourcePath);
-        });
+        let res = await this.utils.deleteModel(model);
+        if (!res) return;
+        this.restFileSystem.removeFile(model.resourcePath);
     }
 
 }
